+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Upcoming...
 
+* New config: 'download_inline' #141 (download configured extensions inline in the browser)
 * Korean language added #119 (Thanks Jinhee-Kim)
 * Galician language added #126 (Thanks vinpoloaire)
 * Russian language added #128 (Thanks BagriyDmitriy)